+++ libs/mb-utils/mb_square_root_matrix.c
@@ -5,7 +5,7 @@ void mb_square_root_matrix(size_t rows, size_t cols, float matrix[rows][cols], f
     for (int j = 0; j < rows; ++j) {
         for (int k = 0; k < cols; ++k) {
             float element = matrix[j][k];
-            sqrt_matrix[j][k] = sqrt(element);
+            sqrt_matrix[j][k] = sqrtf(element);
         }
     }
 }
